+++ dev/tests/node.js
@@ -1,0 +1,6 @@
+const type6 = require("../../dist/type6.cjs");
+
+let vector = new type6.Vector2(2.0,2.2);
+console.log(vector.toString());
+console.log(type6.Utils.round(6.7, 0));
+console.log(type6.Random.float(4, 10));